+++ opentelemetry-stackdriver/CHANGELOG.md
@@ -2,7 +2,12 @@
 
 ## vNext
 
+- Update gRPC schemas
+
 ### Changed
+
+- Added support for `MonitoredResource::CloudFunction`, `MonitoredResource::AppEngine`,
+  `MonitoredResource::ComputeEngine`, and `MonitoredResource::KubernetesEngine`
 
 ## v0.25.0